+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.5.5
+* Connectivity status
+* Replace GetStorage with Hive
+
 ## 0.5.4
 * Added Field<MediaFile> as default gql field